requirements.txt
Flask>=2.3
Werkzeug>=2.3
SQLAlchemy>=2.0
Flask-SQLAlchemy>=3.0
Flask-Login>=0.6
Flask-WTF>=1.1
Flask-SocketIO>=5.3
python-socketio>=5.8
python-engineio>=4.3
mutagen>=1.47
pydub>=0.25
Unidecode>=1.3
pytz
blinker>=1.8
itsdangerous>=2.2
Jinja2>=3.1
MarkupSafe>=2.1
click>=8.1
eventlet
gunicorn
python-dotenv